requirements.txt
# Core Dependencies - Pinned for Production
# Telegram Bot API
python-telegram-bot==20.7

# Configuration
PyYAML==6.0.1

# Security
cryptography==41.0.7

# Local LLM/ML libraries - Pinned versions
torch==2.1.2
transformers==4.35.2
tokenizers==0.15.0

# Vision/Multimodal models
Pillow==10.1.0
opencv-python==4.8.1.78

# HTTP requests
requests==2.31.0
aiohttp==3.9.1

# Logging and utilities
python-dotenv==1.0.0

# Performance optimization (optional)
pyahocorasick==2.0.0  # For efficient keyword matching

# Optional: Popular local LLM frameworks (uncomment as needed)
# llama-cpp-python==0.2.0
# sentence-transformers==2.2.0
# langchain==0.1.0
# ollama==0.1.0

# Web Dashboard dependencies
flask==3.0.3
flask-socketio==5.5.1
flask-cors==6.0.1
plotly==5.24.1

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0